Volume Window s3x.  Auto-displayed song info is disabled right now.  Will thread to eliminate problems soon.  I've noted, however, that I can press keys rapidly to adjust the volume without crashing.

+++ StatusWindowController.h
@@ -47,7 +47,7 @@ typedef enum {
 
 - (void)showUpcomingSongsWithTitles:(NSArray *)titleStrings;
 
-- (void)showVolumeWindowWithLevel:(int)level;
+- (void)showVolumeWindowWithLevel:(float)level;
 - (void)showRatingWindowWithLevel:(int)level;
 - (void)showShuffleWindowWithMode:(MTStatusWindowShuffleMode)mode;
 - (void)showLoopWindowWithMode:(MTStatusWindowLoopMode)mode;
